Improved all administration

+++ app/Api/V1/Controllers/ImageController.php
@@ -63,10 +63,14 @@ class ImageController extends Controller
     }
 
     if(file_exists($request->get('image'))) {
-      if(unlink($request->get('image'))) {
-        return $this->response->noContent();
+      if(!strcmp($request->get('image'), 'images/default/user.jpg')) {
+        if(unlink($request->get('image'))) {
+          return $this->response->noContent();
+        } else {
+          return $this->response->errorInternal('could_not_remove_image');
+        }
       } else {
-        return $this->response->errorInternal('could_not_remove_image');
+        return $this->response->noContent();
       }
     } else {
       return $this->response->noContent();
